Add multi-endpoint broadcast tests
