+++ ui-lib/build.gradle.kts
@@ -216,6 +216,7 @@ dependencies {
 
     testImplementation(libs.kotlin.test)
     testImplementation("io.ktor:ktor-client-mock")
+    testImplementation(libs.kotlinx.coroutines.test)
 
     androidTestImplementation(projects.testLib)
     androidTestImplementation(libs.bundles.androidx.test)
